Airflow3
This commit moves the Docker Compose configuration over to using Docker
Images built by Github Actions in the bluecore_api and
bluecore_workflows repositories. You will need a GitHub token with
permission to read the GitHub Container Registry for these images, and
store it as `CR_PAT` in your `.env` file. I thought making the images
public would mean this isn't needed, but it appears to still be needed
in order to read from GHCR?

In addition to removing the Docker configuration for api and workflows,
the nginx Docker configuration was removed and switched over to use the
image directly, with the config and index.html mounted in the container.