Set internal Airflow API URL
For Airflow to talk back to itself (worker container -> api container) it needs to know the modified base URL that is
being used, so that nginx can proxy requests correctly.

```
AIRFLOW__CORE__EXECUTION_API_SERVER_URL: 'http://airflow-apiserver:8080/workflows/execution/'
```

Prior to settning this we were seeing 404s because the worker was trying to talk to http://airflow-apiserver:8080/execution/ which didn't exist.

I also updated the default keycloak realm to allow the new redirect URI that comes with Airflow 3.